Add files via upload
There is a gene sequence showing activity at the hdlbpa exon on chromosome 3. The downstream loop shows an open chromatin region from ATAC data. The larvae ATAC data shows that the region has a noncoding OCR; whereas the diapause data shows only the open chromatin region. It is possible that this OCR remains closed under diapause, and acts as a gene regulator or transcription factor for larvae not in diapause.